Bug fixes, support for module updates, and better web app support
In addition to a few tweaks/fixes encountered in testing, the Installer has been reworked/expanded to support 'update' operations
besides the 'install' operation - this necessitated changes in DB schema. The installation of web apps was tackled as well, with
OAI-PMH as a particularly simple test case. Note that the servlet dependency has been bumped to 3.0, so Tomcat 7+, Jetty 8+ now
required